Text/LLM/requirements.txt
# pip install -r requirements.txt
# === FastAPI 서버 ===
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
starlette==0.36.3
sse-starlette==1.6.5

# === 비동기 처리 & 웹 ===
aiohttp==3.9.5
aiosignal==1.3.2
anyio==4.9.0
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
frozenlist==1.6.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
hyperframe==6.1.0
idna==3.10
multidict==6.4.3
nest_asyncio==1.6.0
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
urllib3==2.4.0
websockets==14.2
yarl==1.20.0

# === LangChain + 관련 모듈 ===
langchain-core>=0.3.0
langchain>=0.3.25
langchain-community>=0.3.24
langchain-huggingface>=0.0.3
langchain-qdrant>=0.2.0
langchain-text-splitters>=0.3.8
langchainhub>=0.1.21
langgraph>=0.4.3
langgraph-checkpoint>=2.0.25
langgraph-prebuilt>=0.1.8
langgraph-sdk>=0.1.69
langsmith>=0.3.42

# === Hugging Face ===
huggingface-hub==0.31.2
transformers>=4.39.0
tokenizers>=0.19.1
--find-links https://download.pytorch.org/whl/cu121
torch==2.2.1
torchvision==0.17.1

# === Embedding & VectorDB ===
qdrant-client==1.10.1
safetensors==0.5.3
sentence-transformers==2.6.1
xxhash==3.5.0
zstandard==0.23.0

# === Google Cloud 연동 ===
google-api-core==2.24.2
google-api-python-client==2.124.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-aiplatform==1.92.0
google-cloud-bigquery==3.32.0
google-cloud-core==2.4.3
google-cloud-pubsub==2.18.4
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.62.3
grpcio-tools==1.62.3
protobuf==4.25.7

# === 데이터 처리 & 분석 ===
Bottleneck==1.5.0
jsonpatch==1.33
jsonpointer==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
networkx==3.4.2
numexpr==2.10.2
numpy==1.26.3
pandas==2.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.15.3
setuptools==80.7.1
shapely==2.1.0
SQLAlchemy==2.0.41
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
validators==0.35.0
wcwidth==0.2.13

# === 이미지 처리 ===
opencv-python-headless==4.11.0.86
pillow==11.2.1
pillow-avif-plugin==1.5.2

# === 기타 개발 유틸 ===
annotated-types==0.7.0
beautifulsoup4==4.12.2
cachetools==5.5.2
click==8.2.0
dataclasses-json==0.6.7
distro==1.9.0
docstring_parser==0.16
filelock==3.18.0
fsspec==2025.3.2
Jinja2==3.1.6
joblib==1.5.0
mypy_extensions==1.1.0
orjson==3.10.18
ormsgpack==1.9.1
packaging==24.2
portalocker==2.10.1
propcache==0.3.1
proto-plus==1.26.1
types-requests==2.32.0.20250515
sentencepiece==0.2.0
accelerate==0.27.2

bitsandbytes==0.43.1
rq==2.4.0

# === vLLM ===
# vLLM은 torch 설치 후 별도 설치 필요
# vllm

#191 === SQS ===
boto3

